ds2.12

+++ lib/daemon/races.c
@@ -57,7 +57,7 @@ static void create() {
 }
 
     static private void validate() {
-        if( !((int)master()->valid_apply(({ PRIV_ASSIST }))) )
+        if( !(master()->valid_apply(({ PRIV_ASSIST }))) )
             error("Illegal attempt to modify race data");
     }
