+++ lib/verbs/items/pick.c
@@ -41,7 +41,7 @@ mixed can_pick_str_on_obj(string str) {
     if( this_player()->GetParalyzed() ) {
         return "You cannot do anything.";
     }
-    if( (int)this_player()->GetStaminaPoints() < 20 ) {
+    if( this_player()->GetStaminaPoints() < 20 ) {
         return "You are too tired.";
     }
     return this_player()->CanManipulate();
@@ -51,7 +51,7 @@ mixed can_pick_str_on_obj_with_obj(string str) {
     if( this_player()->GetParalyzed() ) {
         return "You cannot do anything.";
     }
-    if( (int)this_player()->GetStaminaPoints() < 30 ) {
+    if( this_player()->GetStaminaPoints() < 30 ) {
         return "You are too tired.";
     }
     return this_player()->CanManipulate();
@@ -64,7 +64,7 @@ mixed do_pick_obj(object ob, string id) {
 mixed do_pick_str_on_obj(string wrd, object ob, mixed *args...) {
     wrd = remove_article(lower_case(args[1]));
     this_player()->eventPrint("You eye the lock for weaknesses.");
-    if( (int)this_player()->GetInCombat() )
+    if( this_player()->GetInCombat() )
         this_player()->SetAttack(0, (: $(ob)->eventPick(this_player(), $(wrd)):),
                 ROUND_OTHER);
     else ob->eventPick(this_player(), wrd);
@@ -75,7 +75,7 @@ mixed do_pick_str_on_obj_with_obj(string wrd, object ob, object tool,
         mixed *args...) {
     wrd = remove_article(lower_case(args[1]));
     this_player()->eventPrint("You eye the lock for weaknesses.");
-    if( (int)this_player()->GetInCombat() )
+    if( this_player()->GetInCombat() )
         this_player()->SetAttack(0, (: $(ob)->eventPick(this_player(), $(wrd),
                         $(tool)) :),ROUND_OTHER);
     else ob->eventPick(this_player(), wrd, tool);
