+++ lib/www/cgi/edit.c
@@ -6,7 +6,7 @@ inherit LIB_DAEMON;
 mapping UploadsMap = ([]);
 
 void validate(){
-    if(!(int)master()->valid_apply(({ "SECURE", "ASSIST" })) &&
+    if(!master()->valid_apply(({ "SECURE", "ASSIST" })) &&
             strsrch(base_name(previous_object()), SOCKET_HTTP)){
         string offender = identify(previous_object(-1));
         debug("edit.c SECURITY VIOLATION: "+offender+" ",get_stack(),"red");
